chore: extract sidebar constants
