+++ src/lib/components/ui/sidebar/constants.ts
@@ -1,0 +1,15 @@
+import LayoutDashboard from 'lucide-svelte/icons/layout-dashboard';
+import WalletCards from 'lucide-svelte/icons/wallet-cards';
+
+export const navLinks = [
+	{
+		title: 'Dashboard',
+		href: 'dashboard',
+		icon: LayoutDashboard
+	},
+	{
+		title: 'Projects',
+		href: 'projects',
+		icon: WalletCards
+	}
+];
